refactor(profiling) Move crashtracker to its own package (#294)
