+++ LICENSE-3rdparty.yml
@@ -1,4 +1,4 @@
-root_name: datadog-profiling, ddcommon, datadog-profiling-ffi, ddcommon-ffi, datadog-profiling-replayer, ddtelemetry, ddtelemetry-ffi, tools, datadog-ipc, datadog-ipc-macros, tarpc, tarpc-plugins, spawn_worker, cc_utils, datadog-sidecar, datadog-sidecar-macros, datadog-trace-normalization, datadog-trace-protobuf, datadog-trace-utils, datadog-sidecar-ffi, sidecar_mockgen, datadog-trace-obfuscation, test_spawn_from_lib, datadog-serverless-trace-mini-agent
+root_name: datadog-crashtracker, datadog-profiling, ddcommon, datadog-profiling-ffi, ddcommon-ffi, datadog-profiling-replayer, ddtelemetry, ddtelemetry-ffi, tools, datadog-ipc, datadog-ipc-macros, tarpc, tarpc-plugins, spawn_worker, cc_utils, datadog-sidecar, datadog-sidecar-macros, datadog-trace-normalization, datadog-trace-protobuf, datadog-trace-utils, datadog-sidecar-ffi, sidecar_mockgen, datadog-trace-obfuscation, test_spawn_from_lib, datadog-serverless-trace-mini-agent
 third_party_libraries:
 - package_name: addr2line
   package_version: 0.21.0
