+++ crashtracker/src/collectors.rs
